Changed widescreen FOV formula to Hor+
Previously the widescreen support code was using another formula. While that
does formula does work for 90 degree FOV, it gives slightly incorrect results
for other FOV values.

Hor+ is the formula most commonly used in games. In that formula the vertical
FOV stays the same while the horizontal FOV is expanded to fit the screen.
Switching to this formula makes the FOV values consistent with that is used in
most other games.